code lẹ chớ rớt môn :(

+++ shop_c/webs/src/main/webapp/js/cart.js
@@ -22,7 +22,6 @@ function updateTotal(input) {
 	updateCartTotal(cartId, total, quantity);
 }
 
-// Hàm gọi API cập nhật tổng tiền (nếu cần thiết)
 function updateCartTotal(idCart, total, quantity) {
 	fetch(`cart/upd.htm?idCart=${idCart}&quantity=${quantity}&total=${total}`, {
 		method: 'GET',
@@ -109,6 +108,9 @@ window.onload = function() {
 function deleteCart(idCart) {
 	// Hiển thị thông báo xác nhận
 	const isConfirmed = window.confirm("Bạn có chắc chắn muốn xóa sản phẩm này?");
+	const oldQuantityCartText = document.querySelector('#quantityCart').innerHTML.trim();
+	const oldQuantityCart = parseInt(oldQuantityCartText, 10) || 0;
+	const newQuantityCart = oldQuantityCart - 1;
 
 	// Nếu người dùng chọn "OK" (true), tiếp tục thực hiện xóa
 	if (isConfirmed) {
@@ -121,8 +123,13 @@ function deleteCart(idCart) {
 			.then(response => response.json())  // Nếu bạn trả về JSON từ server
 			.then(data => {
 				console.log('Xóa sản phẩm thành công!!!', data);
-				// Sau khi xóa thành công, gọi hàm để xóa phần tử HTML
-				deleteElementById(idCart);  // Xóa phần tử HTML tương ứng với idCart
+				const allCartProducts = document.querySelectorAll('.cartProduct');
+				if (allCartProducts.length === 1) {
+					window.location.href = 'cart.htm';
+				} else {
+					deleteElementById(idCart);  // Xóa phần tử HTML tương ứng với idCart
+					document.querySelector('#quantityCart').innerHTML = newQuantityCart;
+				}
 			})
 			.catch(error => console.error('Lỗi:', error));
 	} else {
